+++ app/src/main/java/org/ssg_tab/data/remote/datasource/quiz/QuizDataSource.kt
@@ -1,6 +1,6 @@
-package org.ssg_tab.data.remote.datasource
+package org.ssg_tab.data.remote.datasource.quiz
 
-import org.ssg_tab.data.dto.response.BaseResponse
+import org.ssg_tab.data.dto.response.base.BaseResponse
 import org.ssg_tab.data.dto.response.quiz.QuizListResponseDto
 
 interface QuizDataSource {
